fix(backend/mmdet3d): support export to onnx
