+++ mmdeploy/codebase/mmdet3d/deploy/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .mmdetection3d import MMDetection3d
-from .voxel_detection import VoxelDetection
+from .mmdetection3d import MMDetection3d, VoxelDetection
+# from .voxel_detection import 
 from .voxel_detection_model import VoxelDetectionModel
 
 __all__ = ['MMDetection3d', 'VoxelDetection', 'VoxelDetectionModel']
