+++ tools/check_env.py
@@ -1,7 +1,7 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 
 from mmcv.utils import collect_env as collect_base_env
-from mmcv.utils import get_git_hash
+from mmengine import get_git_hash
 
 import mmdeploy
 from mmdeploy.utils import (get_backend_version, get_codebase_version,
